Hoist YouTube component et al out of it's own folder
