+++ src/YoutubeAtom.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
 import { render } from '@testing-library/react';
-import { Pillar, Display, Design } from '@guardian/types/Format';
 import { YoutubeAtom } from './YoutubeAtom';
 
 describe('YoutubeAtom', () => {
@@ -9,11 +8,6 @@ describe('YoutubeAtom', () => {
         const atom = (
             <YoutubeAtom
                 title="My Youtube video!"
-                format={{
-                    pillar: Pillar.News,
-                    design: Design.Article,
-                    display: Display.Standard,
-                }}
                 videoMeta={{
                     assetId: '-ZCvZmYlQD8',
                     mediaTitle: 'YoutubeAtom',
