+++ src/YoutubeAtom.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
+import { css } from 'emotion';
 
 import { YoutubeOverlay } from './YoutubeOverlay';
-import { MaintainAspectRatio } from './MaintainAspectRatio';
 
 type EmbedConfig = {
     adsConfig: {
@@ -63,6 +63,35 @@ type YoutubeAtomType = {
     duration?: number; // in seconds
     origin?: string;
 };
+
+const MaintainAspectRatio = ({
+    height,
+    width,
+    children,
+}: {
+    height: number;
+    width: number;
+    children: React.ReactNode;
+}): JSX.Element => (
+    /* https://css-tricks.com/aspect-ratio-boxes/ */
+    <div
+        className={css`
+            /* position relative to contain the absolutely positioned iframe plus any Overlay image */
+            position: relative;
+            padding-bottom: ${(height / width) * 100}%;
+
+            iframe {
+                width: 100%;
+                height: 100%;
+                position: absolute;
+                top: 0;
+                left: 0;
+            }
+        `}
+    >
+        {children}
+    </div>
+);
 
 // Note, this is a subset of the CAPI MediaAtom essentially.
 export const YoutubeAtom = ({
